% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_to_aa}
\alias{codon_to_aa}
\title{Title}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codon_to_aa
}
\description{
A named vector containing codons and the amino acids they are translated to
}
\keyword{datasets}
