% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Transcribe DNA to RNA}
\usage{
dna_to_rna(dna)
}
\arguments{
\item{dna}{dna sequence which gets transcribed to rna}
}
\value{
transcribed rna
}
\description{
Transcribe DNA to RNA
}
\examples{
dna_to_rna("GTGA")
}
