% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aa_from_codons.R
\name{get_aa_from_codon}
\alias{get_aa_from_codon}
\title{Get amino acids from a list of codons}
\usage{
get_aa_from_codon(codons)
}
\arguments{
\item{codons}{a vector of codons}
}
\value{
a string of a amino acids
}
\description{
Get amino acids from a list of codons
}
