% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codons.R
\name{get_codons}
\alias{get_codons}
\title{Find the codons in a given DNA string}
\usage{
get_codons(dna_str, start = 1)
}
\arguments{
\item{dna_str}{a string of nucleotides}

\item{start}{the position of the read frame, default=1}
}
\value{
vector of codons
}
\description{
Find the codons in a given DNA string
}
