% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DNA.R
\name{make_DNA}
\alias{make_DNA}
\title{Generate a random string of nucleotides}
\usage{
make_DNA(nuc_len)
}
\arguments{
\item{nuc_len}{number of nucleotides to output}
}
\value{
a string of nucleotides of length nuc_len
}
\description{
Generate a random string of nucleotides
}
