% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seq_composition.R
\name{plot_seq_composition}
\alias{plot_seq_composition}
\title{Plotting the sequence composition}
\usage{
plot_seq_composition(sequence)
}
\arguments{
\item{Sequence}{Input of the function is the sequence you would like to analyze}
}
\value{
It will output a plot with the character counts
}
\description{
Plotting the sequence composition
}
\examples{
plot_seq_composition("AGTGTAA")
}
